Skeleton complete for NoData and NoAccess. Transaction sync
All transactions now have a local and remote ID, to allow
synchronisation between the two peers. When a data request is sent an id
is generated for the requestor. When the provider receives it, they
generate their own id and store both. They can now lookup the
requestor's ID when communicating about this transaction