Fixed batch generator, attempted to optimise messages
Batch Generator was appending JVM arguments to the wrong place, so heap
size was not being set. This is now fixed. Messages (the thing we have
literally millions of) were storing a full copy of the sender and
receiver Node, hence the huge memory foot print (in theory). They now
send just the ID, and the receiving peer can look up the actual Node
dynamically.